student managing page complete with teacher bug fixed

+++ public/includes/class/delete_user_role.php
@@ -52,7 +52,10 @@ class OE_Delete_user
     }
     public static function delete_post($user_id)
     {
-        wp_delete_post($user_id, true);
+        if (get_permalink($user_id)) {
+            wp_delete_post($user_id, true);
+        }
+
     }
     public static function delete_students($user_id)
     {
